src/resources/requirement.txt
# Data sources
# News articles
GoogleNews==1.6.15
newsapi-python==0.2.7

# Financial reports
sec-edgar-downloader==5.0.3

# Fundamental data
yfinance==0.2.65
finnhub-python==2.4.24
simfin==1.0.1

# Social media
tweepy==4.16.0
praw==7.8.1

# Market data
#yfinance==0.2.65
